Make ASGs not fail parsing when new fields are added
Isn't the _best_ solution but we track our depth as we traverse the
fields and use that information to close an ASG record. We do this now
instead of assuming assuming that an `end_element` of "member" when we're
not within an explicitly tracked context (The `@in_<element name>` ivars)
means we have finished.